+++ examples/suttmeier.py
@@ -12,7 +12,7 @@ from viamr import VIAMR
 
 print = PETSc.Sys.Print  # enables correct printing in parallel
 
-refinements = 6
+refinements = 4
 m0 = 10
 outfile = "result_suttmeier.pvd"
 
@@ -32,8 +32,10 @@ params = {
     "pc_factor_mat_solver_type": "mumps",
 }
 
+# explicitly setting distribution parameters allows this to be a udomark() example
+# which still runs in parallel
 meshhierarchy = [
-    UnitSquareMesh(m0, m0),
+    UnitSquareMesh(m0, m0, diagonal="crossed", distribution_parameters={"partition": True, "overlap_type": (DistributedMeshOverlapType.VERTEX, 1)}),
 ]
 amr = VIAMR()
 for i in range(refinements + 1):
@@ -70,10 +72,12 @@ for i in range(refinements + 1):
     if i == refinements:
         break
 
-    # apply VCD AMR, marking inactive by B&R indicator
+    mark = amr.udomark(u, psi, n = 2)
+
+    # alternative: apply VCD AMR, marking inactive by B&R indicator
     #   (choose more refinement in active set, relative to default bracket=[0.2, 0.8])
-    mark = amr.vcdmark(u, psi, bracket=[0.1, 0.8])
-    #mark = amr.udomark(u, psi, n = 2)
+    # mark = amr.vcdmark(u, psi, bracket=[0.1, 0.8])
+
     residual = -div(grad(u)) - fsource
     (imark, _, _) = amr.brinactivemark(u, psi, residual)
     # imark = amr.eleminactive(u, psi)  # alternative is to refine all inactive
