+++ tests/3drefine.py
@@ -8,9 +8,8 @@ amr = VIAMR()
 mesh = BoxMesh(5, 5, 10, 1, 1, 1)
 DG0 = FunctionSpace(mesh, "DG", 0)
 (x, y, z) = SpatialCoordinate(mesh)
-indicator = Function(DG0).interpolate(conditional(x>.5, 1, 0))
+indicator = Function(DG0).interpolate(conditional(x > 0.5, 1, 0))
 
 refinedmesh = amr.refinemarkedelements(mesh, indicator)
 
 VTKFile("3d.pvd").write(refinedmesh)
-
