+++ tests/compareSolution.py
@@ -7,17 +7,29 @@ import subprocess
 z = VIAMR()
 
 
-subprocess.run(["python3", "generateSolution.py", "--refinements", "2", "--runtime", "serial"])
 subprocess.run(
-    ["mpiexec", "-n", "4", "python3", "generateSolution.py",
-     "--refinements", "2", "--runtime", "parallel"])
+    ["python3", "generateSolution.py", "--refinements", "2", "--runtime", "serial"]
+)
+subprocess.run(
+    [
+        "mpiexec",
+        "-n",
+        "4",
+        "python3",
+        "generateSolution.py",
+        "--refinements",
+        "2",
+        "--runtime",
+        "parallel",
+    ]
+)
 
-with CheckpointFile("serialUDO.h5", 'r') as afile:
-    serialMesh = afile.load_mesh("serialMesh") 
+with CheckpointFile("serialUDO.h5", "r") as afile:
+    serialMesh = afile.load_mesh("serialMesh")
     serialMark = afile.load_function(serialMesh, "serialMark")
-    
-    
-with CheckpointFile("parallelUDO.h5", 'r') as afile:
+
+
+with CheckpointFile("parallelUDO.h5", "r") as afile:
     parallelMesh = afile.load_mesh("parallelMesh")
     parallelMark = afile.load_function(parallelMesh, "parallelMark")
